pyproject.toml
[tool.pdm]

[project]
name = "elephant_news"
version = "0.1.0"
description = "Non-biased AI-summarized news from primary sources."
authors = [
    {name = "Matt Zhang", email = "set.stun@gmail.com"},
]
dependencies = [
    "openai>=0.27.6",
    "prompt-toolkit>=3.0.38",
    "python-dotenv>=1.0.0",
    "termcolor>=2.3.0",
    "pathlib>=1.0.1",
    "tiktoken>=0.4.0",
    "dataclasses-json>=0.5.7",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"
